challenges/problem-dev/requirements.txt
fastapi
uvicorn
pycryptodome
pycryptodomex
itsdangerous
